REFACTOR: Star, StaticStar, ShootingStar 컴포넌트 common=>chracter로 경로 이동
