+++ src/components/character/Star.tsx
