+++ src/components/character/StaticStar.tsx
